// src/extended/cores/types.ts
export interface ListParam {
    /**
     * max items count to be fetched
     */
    limit?: number;
    /**
     * (optional) sorting order
     *  - 'asc': older first
     *  - 'desc': newer first
     *  - string: extended sorting features
     */
    sort?: 'asc' | 'desc' | string;
}

/**
 * type `PaginateParam`
 */
export interface PaginateParam extends ListParam {
    /**
     * page # to fetch (0-indexed)
     */
    page?: number;
    /**
     * (optional) offset # from start.
     */
    offset?: number;
    /**
     * (optional) flag to filter by uid
     */
    uid?: string;
    /**
     * (optional) flag to filter by sid (only for admin role)
     */
    sid?: string;
}